@@ -1,6 +1,6 @@
 {
   "private": true,
-  "version": "2020.8.0",
+  "version": "2020.8.1",
   "workspaces": [
     "packages/*",
     "examples/*",
@@ -12,30 +12,30 @@
     }
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^3.5.0",
-    "@typescript-eslint/parser": "^3.5.0",
+    "@typescript-eslint/eslint-plugin": "^3.8.0",
+    "@typescript-eslint/parser": "^3.8.0",
     "babel-eslint": "10.1.0",
     "cz-conventional-changelog": "^3.1.0",
-    "eslint": "^7.3.1",
+    "eslint": "^7.6.0",
     "eslint-config-prettier": "^6.10.1",
     "eslint-config-react-app": "^5.2.1",
     "eslint-plugin-cypress": "^2.11.1",
     "eslint-plugin-flowtype": "5.2.0",
     "eslint-plugin-import": "2.22.0",
-    "eslint-plugin-jsdoc": "^28.6.1",
+    "eslint-plugin-jsdoc": "^30.2.1",
     "eslint-plugin-jsx-a11y": "6.3.1",
     "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.20.3",
-    "eslint-plugin-react-hooks": "^4.0.5",
+    "eslint-plugin-react": "^7.20.5",
+    "eslint-plugin-react-hooks": "^4.0.8",
     "husky": "^4.2.3",
     "isomorphic-fetch": "^2.2.1",
     "jest-environment-jsdom-sixteen": "^1.0.3",
     "lerna": "^3.22.1",
     "lint-staged": "^10.2.11",
     "prettier": "^2.0.4",
-    "start-server-and-test": "^1.10.11",
-    "typescript": "^3.9.5",
-    "wait-on": "^5.0.1"
+    "start-server-and-test": "^1.11.2",
+    "typescript": "^3.9.7",
+    "wait-on": "^5.1.0"
   },
   "scripts": {
     "cy:run:dms": "cypress run --project ./examples/sn-dms-demo",
