@@ -2,94 +2,33 @@
  * @module ViewControls
  */
 import { isExtendedError } from '@sensenet/client-core'
-import { FieldSetting, GenericContent } from '@sensenet/default-content-types'
+import { EditView as SnEditView } from '@sensenet/controls-react'
+import { GenericContent } from '@sensenet/default-content-types'
 import { useLogger, useRepository } from '@sensenet/hooks-react'
-import { createStyles, makeStyles } from '@material-ui/core'
-import Button from '@material-ui/core/Button'
-import Grid from '@material-ui/core/Grid'
-import clsx from 'clsx'
-import React, { createElement, ReactElement, useEffect, useState } from 'react'
-import MediaQuery from 'react-responsive'
+import React, { ReactElement, useEffect, useState } from 'react'
 import { useHistory, useRouteMatch } from 'react-router-dom'
-import { globals, useGlobalStyles } from '../../globalStyles'
+import { useGlobalStyles } from '../../globalStyles'
 import { useLocalization, useSelectionService } from '../../hooks'
 import { navigateToAction } from '../../services'
 import { ActionNameType, reactControlMapper } from '../react-control-mapper'
-import { ViewTitle } from './view-title'
+import { useViewControlStyles } from './common/styles'
+import { ViewTitle } from './common/view-title'
 
-const useStyles = makeStyles(() => {
-  return createStyles({
-    form: {
-      margin: '0 auto',
-      padding: '22px 22px 0 22px',
-      overflowY: 'auto',
-      width: '100%',
-      height: `calc(100% - ${globals.common.formActionButtonsHeight}px)`,
-    },
-    formFullPage: {
-      height: `calc(100% - ${globals.common.formActionButtonsHeight + globals.common.drawerItemHeight}px)`,
-    },
-    mainForm: {
-      display: 'initial',
-      height: `calc(100% - ${globals.common.formTitleHeight}px)`,
-    },
-    mainFormFullpage: {
-      height: `calc(100% - ${globals.common.drawerItemHeight}px)`,
-    },
-    grid: {
-      display: 'flex',
-      alignItems: 'center',
-      flexFlow: 'column',
-      padding: '15px !important',
-      height: 'fit-content',
-      position: 'relative',
-    },
-    wrapper: {
-      width: '75%',
-      position: 'relative',
-    },
-    wrapperFullWidth: {
-      width: '88%',
-    },
-    actionButtonWrapper: {
-      height: '80px',
-      width: '100%',
-      position: 'absolute',
-      padding: '20px',
-      bottom: 0,
-      textAlign: 'right',
-    },
-  })
-})
-
-/**
- * Interface for EditView properties
- */
 export interface EditViewProps {
   renderIcon?: (name: string) => ReactElement
   handleCancel?: () => void
   submitCallback?: () => void
-  uploadFolderpath?: string
   actionName?: ActionNameType
   isFullPage?: boolean
   contentPath: string
 }
 
-/**
- * View Control for editing a Content, works with a single Content and based on the ReactControlMapper
- *
- * Usage:
- * ```html
- *  <EditView contentPath={selectedContentPath} submitCallback={editSubmitClick} />
- * ```
- */
 export const EditView: React.FC<EditViewProps> = (props) => {
-  const repo = useRepository()
+  const repository = useRepository()
   const selectionService = useSelectionService()
   const [content, setContent] = useState<GenericContent>()
-  const controlMapper = reactControlMapper(repo)
-  const [saveableFields, setSaveableFields] = useState({})
-  const classes = useStyles()
+  const controlMapper = reactControlMapper(repository)
+  const classes = useViewControlStyles()
   const globalClasses = useGlobalStyles()
   const localization = useLocalization()
   const logger = useLogger('EditView')
@@ -98,7 +37,7 @@ export const EditView: React.FC<EditViewProps> = (props) => {
 
   useEffect(() => {
     async function getExpandedContent() {
-      const expanedContentResponse = await repo.load({
+      const expanedContentResponse = await repository.load({
         idOrPath: props.contentPath,
         oDataOptions: {
           select: 'all',
@@ -109,144 +48,74 @@ export const EditView: React.FC<EditViewProps> = (props) => {
       selectionService.activeContent.setValue(expanedContentResponse.d)
     }
     getExpandedContent()
-  }, [repo, props.contentPath, selectionService.activeContent])
+  }, [repository, props.contentPath, selectionService.activeContent])
 
   if (content === undefined) {
     return null
   } else {
-    const schema = controlMapper.getFullSchemaForContentType(
-      content.Type,
-      props.actionName ? props.actionName : 'browse',
-    )
-
-    const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
-      event.preventDefault()
+    const handleSubmit = async (saveableFields: GenericContent) => {
       try {
-        const response = repo.patch({
+        const response = repository.patch({
           idOrPath: content.Id,
           content: saveableFields,
         })
         logger.information({
           message: localization.editPropertiesDialog.saveSuccessNotification.replace(
             '{0}',
-            content.DisplayName || content.Name || content.DisplayName || content.Name,
+            saveableFields.DisplayName || saveableFields.Name || content.DisplayName || content.Name,
           ),
           data: {
             relatedContent: content,
             content: response,
-            relatedRepository: repo.configuration.repositoryUrl,
+            relatedRepository: repository.configuration.repositoryUrl,
           },
         })
       } catch (error) {
         logger.error({
           message: localization.editPropertiesDialog.saveFailedNotification.replace(
             '{0}',
-            content.DisplayName || content.Name || content.DisplayName || content.Name,
+            saveableFields.DisplayName || saveableFields.Name || content.DisplayName || content.Name,
           ),
           data: {
             relatedContent: content,
             content,
-            relatedRepository: repo.configuration.repositoryUrl,
-            error: isExtendedError(error) ? repo.getErrorFromResponse(error.response) : error,
+            relatedRepository: repository.configuration.repositoryUrl,
+            error: isExtendedError(error) ? repository.getErrorFromResponse(error.response) : error,
           },
         })
       } finally {
-        props.submitCallback && props.submitCallback()
+        props.submitCallback?.()
       }
     }
 
-    const handleInputChange = (field: string, value: unknown) => {
-      setSaveableFields({ ...saveableFields, [field]: value })
-    }
-
-    const isFullWidthField = (field: { fieldSettings: FieldSetting }) => {
-      return (
-        (field.fieldSettings.Name === 'Avatar' && content.Type.includes('User')) ||
-        (field.fieldSettings.Name === 'Enabled' && content.Type.includes('User')) ||
-        field.fieldSettings.Type === 'LongTextFieldSetting'
-      )
-    }
-
     return (
-      <>
-        <ViewTitle
-          title={props.actionName === 'browse' ? 'Info about' : 'Edit'}
-          titleBold={content?.DisplayName}
-          content={content}
-        />
-        <form
-          className={clsx(classes.mainForm, {
-            [classes.mainFormFullpage]: props.isFullPage,
-          })}
-          onSubmit={handleSubmit}>
-          <div
-            className={clsx(classes.form, {
-              [classes.formFullPage]: props.isFullPage,
-            })}>
-            <Grid container={true} spacing={2}>
-              {schema.fieldMappings
-                .sort((item1, item2) => (item2.fieldSettings.FieldIndex || 0) - (item1.fieldSettings.FieldIndex || 0))
-                .map((field) => {
-                  const fieldControl = createElement(
-                    controlMapper.getControlForContentField(
-                      content.Type,
-                      field.fieldSettings.Name,
-                      props.actionName ? props.actionName : 'browse',
-                    ),
-                    {
-                      repository: repo,
-                      settings: field.fieldSettings,
-                      content,
-                      fieldValue: (content as any)[field.fieldSettings.Name],
-                      actionName: props.actionName,
-                      renderIcon: props.renderIcon,
-                      fieldOnChange: handleInputChange,
-                      uploadFolderPath: props.uploadFolderpath,
-                    },
-                  )
-
-                  return (
-                    <Grid
-                      item={true}
-                      xs={12}
-                      sm={12}
-                      md={isFullWidthField(field) ? 12 : 6}
-                      lg={isFullWidthField(field) ? 12 : 6}
-                      xl={isFullWidthField(field) ? 12 : 6}
-                      key={field.fieldSettings.Name + content.Id}
-                      className={classes.grid}>
-                      <div
-                        className={clsx(classes.wrapper, {
-                          [classes.wrapperFullWidth]: isFullWidthField(field),
-                        })}>
-                        {fieldControl}
-                      </div>
-                    </Grid>
-                  )
-                })}
-            </Grid>
-          </div>
-          <div className={classes.actionButtonWrapper}>
-            <MediaQuery minDeviceWidth={700}>
-              <Button
-                aria-label={localization.forms.cancel}
-                color="default"
-                className={globalClasses.cancelButton}
-                onClick={async () => {
-                  navigateToAction({ history, routeMatch })
-                  props.handleCancel?.()
-                }}>
-                {localization.forms.cancel}
-              </Button>
-            </MediaQuery>
-            {props.actionName !== 'browse' && (
-              <Button aria-label={localization.forms.submit} variant="contained" color="primary" type="submit">
-                {localization.forms.submit}
-              </Button>
-            )}
-          </div>
-        </form>
-      </>
+      <SnEditView
+        actionName={props.actionName}
+        content={content}
+        onSubmit={handleSubmit}
+        repository={repository}
+        contentTypeName={content.Type}
+        handleCancel={() => navigateToAction({ history, routeMatch })}
+        showTitle={true}
+        uploadFolderpath="/Root/Content/demoavatars"
+        controlMapper={controlMapper}
+        localization={{ submit: localization.forms.submit, cancel: localization.forms.cancel }}
+        classes={{
+          grid: classes.grid,
+          fieldWrapper: classes.fieldWrapper,
+          field: classes.field,
+          fieldFullWidth: classes.fieldFullWidth,
+          actionButtonWrapper: classes.actionButtonWrapper,
+          cancel: globalClasses.cancelButton,
+        }}
+        renderTitle={() => (
+          <ViewTitle
+            title={props.actionName === 'browse' ? 'Info about' : 'Edit'}
+            titleBold={content?.DisplayName}
+            content={content}
+          />
+        )}
+      />
     )
   }
 }
