@@ -1,14 +1,4 @@
 const values = {
-  dashboard: {
-    errorLoadingWidget: 'Error loading widget :(',
-    refresh: 'Refresh',
-    openInSearch: 'Open in Search',
-    updates: {
-      title: 'Packages to update',
-      allUpToDate: 'All packages are up to date',
-      view: 'View',
-    },
-  },
   addButton: {
     tooltip: 'Create or upload content',
     new: 'New...',
@@ -37,6 +27,7 @@ const values = {
       executePrimaryText: '{0} 👉 {1}',
       executeSecondaryText: "Execute custom action '{1}' on content '{0}'",
     },
+    searchSuggestionList: 'Search suggestion list',
   },
   lockedCell: {
     checkedOutTo: (name: string) => `Checked out to ${name}`,
@@ -120,35 +111,63 @@ const values = {
     titles: {
       Content: 'Content',
       ContentTypes: 'Content Types',
-      Dashboard: 'Dashboard',
       Localization: 'Localization',
       Search: 'Search',
       Setup: 'Setup',
       Trash: 'Trash',
-      Query: 'Query',
       UsersAndGroups: 'Users and groups',
       CustomContent: 'Custom content',
     },
     descriptions: {
       Content: 'Explore and manage your content in the repository',
       ContentTypes: 'Manage content types',
-      Dashboard: 'Repository overview',
       Localization: 'Manage string resources',
       Search: 'Execute custom searches, build and save queries',
       Setup: 'Configure the sensenet system',
       Trash: 'Manage deleted items here: restore content or purge them permanently',
-      Query: 'Open my custom Query',
       UsersAndGroups: 'Manage users and groups, roles and identities',
       CustomContent: 'Explore and manage your content from the configured path',
     },
     personalSettingsTitle: 'Edit personal settings',
     personalSettingsSecondaryText: 'Customize the application behavior',
     contentRootDescription: 'The root path. Content will be displayed below this level.',
-    queryTerm: 'Content query see https://community.sensenet.com/docs/content-query/',
     columns: 'Array of columns to display',
     expand: 'Expand',
     collapse: 'Collapse',
     newSearch: 'New search',
+    add: 'Add',
+  },
+  dashboard: {
+    title: (projectName: string) => `Welcome to your ${projectName} project`,
+    descriptionFirstLine: 'Here you can Explore the Admin UI with sample content',
+    descriptionSecondLine: 'Feel free to look around!',
+    subscriptionPlan: 'Your subscription plan',
+    free: 'Free',
+    features: 'Features',
+    users: 'users',
+    content: 'content',
+    storageSpace: 'GB storage space',
+    version: 'Version number',
+    releaseNotes: 'View release notes',
+    getMore: 'Want to get more?',
+    upgrade: 'Upgrade',
+    usage: 'Current usage',
+    Users: 'Users',
+    Content: 'Content',
+    StorageSpace: 'Storage space',
+    used: (current: string | number, limit: string | number) => `${current} of ${limit} used`,
+    yourProject: 'Your project',
+    getStarted: 'Get started with your new project.',
+    learnMore: 'Learn more about Sensenet',
+    learnBasics: 'Learn the basics',
+    learnBasicsDescription: 'Get step-by-step guides and learn what you can achieve only working on the admin-ui.',
+    beExpert: 'Be a content management expert',
+    beExpertDescription: 'Master the concept of sensenet with its special terms and abstractions.',
+    buildApp: 'Build your first app',
+    buildAppDescription: 'Let us help you to begin your sensenet journey!',
+    viewUserGuides: 'View User Guides',
+    viewConceptDocs: 'View Concept Docs',
+    viewDevManual: 'View the Developer manual',
   },
   trash: {
     title: 'Trash',
@@ -217,7 +236,6 @@ const values = {
     drawerEnable: 'Enable or disable the drawer',
     drawerType: 'Drawer type',
     drawerItems: 'Items enabled on the drawer',
-    drawerDashboardName: 'A unique identifier for the dashboard',
     drawerCustomContentAppPath: 'Unique url param for the new menu item',
     drawerCustomContentRoot: 'SenseNet path of the root item of the custom content view',
     drawerItemTitle: 'Title of the item',
@@ -243,28 +261,6 @@ const values = {
     uploadHandlerTitle: 'Select which handlertype(s) has/have upload permission',
     eventLogSize: 'Number of entries to store in the Event Log',
     sendLogWithCrashReports: 'Send log data with crash reports by default',
-    dashboard: {
-      widgetName: 'Widget',
-      minWidth: 'The minimum width of the widget in pixels',
-      widgetType: 'Type of the widget',
-      title: 'Widget title',
-      queryWidget: {
-        settings: 'Settings for the Query widget',
-        query: 'The content query expression',
-        emptyPlaceholderText: 'The text that will be displayed if the query has no hits',
-        showColumnNames: 'Show or hide column names',
-        top: 'Limits the number of hits',
-        showOpenInSearch: 'Option for a button to open the query in the Search view',
-        showRefresh: 'Option for a refresh button',
-        enableSelection: 'Enable content selection',
-        countOnly: 'Displays only the hits count instead of a content list',
-        columns: 'Array of columns to display',
-      },
-      markdownWidget: {
-        settings: 'Settings for the Markdown Widget',
-        content: 'The Markdown content to display',
-      },
-    },
   },
   repositorySelector: {
     loggedInAs: 'You are currently logged in as {0}',
@@ -282,6 +278,7 @@ const values = {
   topMenu: {
     personalSettings: 'Personal settings',
     logout: 'Log out',
+    openMenu: 'Open menu',
   },
   navigationCommandProvider: {
     personalSettingsPrimary: 'Personal Settings',
@@ -448,6 +445,18 @@ const values = {
   },
   contentList: {
     errorContentModification: 'There was an error during content modification',
+  },
+  restore: {
+    description: (contentName: string) =>
+      `You are about to restore <strong>${contentName}</strong> from the Trash to the following destination:`,
+    selectTarget: 'Select target',
+    title: 'Restore',
+    cancel: 'Cancel',
+  },
+  contentPickerDialog: {
+    title: 'Select a target',
+    cancelButton: 'Cancel',
+    selectButton: 'Select',
   },
 }
 
