@@ -19,40 +19,7 @@ export interface UiSettings {
   }
 }
 
-export const widgetTypes = tuple('markdown', 'query', 'updates')
-export interface Widget<T> {
-  title: string
-  widgetType: typeof widgetTypes[number]
-  settings: T
-  minWidth?: PlatformDependent<number | string>
-}
-
-export interface MarkdownWidget extends Widget<{ content: string }> {
-  widgetType: 'markdown'
-}
-
-export interface UpdatesWidget extends Widget<undefined> {
-  widgetType: 'updates'
-}
-
-export interface QueryWidget<T extends GenericContent>
-  extends Widget<{
-    columns: Array<keyof T>
-    showColumnNames: boolean
-    showRefresh?: boolean
-    showOpenInSearch?: boolean
-    enableSelection?: boolean
-    top?: number
-    query: string
-    emptyPlaceholderText?: string
-    countOnly?: boolean
-  }> {
-  widgetType: 'query'
-}
-
-export type WidgetSection = MarkdownWidget | QueryWidget<GenericContent> | UpdatesWidget
-
-export const CustomDrawerItemType = tuple('CustomContent', 'Query', 'Dashboard')
+export const CustomDrawerItemType = tuple('CustomContent')
 
 export const BuiltInDrawerItemType = tuple(
   'Content',
@@ -115,32 +82,7 @@ export interface CustomContentDrawerItem
   itemType: 'CustomContent'
 }
 
-export interface QueryDrawerItem
-  extends DrawerItem<{
-    title: string
-    description?: string
-    icon: string
-    term: string
-    columns: Array<keyof GenericContent>
-  }> {
-  itemType: 'Query'
-}
-
-export interface DashboardDrawerItem
-  extends DrawerItem<{
-    dashboardName: string
-    title: string
-    description?: string
-    icon: string
-  }> {
-  itemType: 'Dashboard'
-}
-
 export type PersonalSettingsType = PlatformDependent<UiSettings> & {
-  dashboards: {
-    globalDefault: WidgetSection[]
-    repositoryDefault: WidgetSection[]
-  } & { [key: string]: WidgetSection[] }
   eventLogSize: number
   sendLogWithCrashReports: boolean
   logLevel: Array<keyof typeof LogLevel>
@@ -150,136 +92,6 @@ export type PersonalSettingsType = PlatformDependent<UiSettings> & {
 }
 
 export const defaultSettings: PersonalSettingsType = {
-  dashboards: {
-    globalDefault: [
-      {
-        title: 'Welcome back, {currentUserName}',
-        widgetType: 'markdown',
-        settings: {
-          content: "It's a great day to do admin stuff!",
-        },
-      },
-    ],
-    repositoryDefault: [
-      {
-        title: 'Welcome back, {currentUserName}',
-        widgetType: 'markdown',
-        settings: {
-          content: "It's a great day to do admin stuff!",
-        },
-        minWidth: {
-          default: '100%',
-        },
-      },
-      {
-        title: 'Packages to update',
-        widgetType: 'updates',
-        minWidth: {
-          default: '100%',
-        },
-        settings: undefined,
-      },
-      {
-        title: 'Number of users',
-        widgetType: 'query',
-        minWidth: { default: '30%' },
-        settings: {
-          query: "+(TypeIs:'User' AND InTree:'/Root/IMS/Public')",
-          columns: [],
-          countOnly: true,
-          showColumnNames: false,
-          showOpenInSearch: false,
-          showRefresh: false,
-        },
-      },
-      {
-        title: 'Number of content items',
-        widgetType: 'query',
-        minWidth: { default: '30%' },
-        settings: {
-          query: "+TypeIs:'GenericContent'",
-          columns: [],
-          countOnly: true,
-          showColumnNames: false,
-        },
-      },
-      {
-        title: 'Updates since yesterday',
-        widgetType: 'query',
-        minWidth: {
-          default: '30%',
-        },
-        settings: {
-          query: '+ModificationDate:>@@Yesterday@@',
-          columns: [],
-          countOnly: true,
-          showColumnNames: false,
-          showOpenInSearch: true,
-        },
-      },
-      {
-        title: 'Docs owned by me',
-        widgetType: 'query',
-        minWidth: {
-          default: '100%',
-        },
-        settings: {
-          query: "+(Owner:@@CurrentUser@@ AND TypeIs:'File')",
-          columns: [],
-          countOnly: true,
-          showColumnNames: false,
-          showOpenInSearch: true,
-        },
-      },
-      {
-        title: 'Docs shared with me',
-        widgetType: 'query',
-        minWidth: {
-          default: '100%',
-        },
-        settings: {
-          query: '+SharedWith:@@CurrentUser@@',
-          columns: [],
-          countOnly: true,
-          showColumnNames: false,
-          showOpenInSearch: true,
-        },
-      },
-      {
-        title: 'Concepts & Tutorials',
-        widgetType: 'markdown',
-        settings: {
-          content:
-            '<div>To get started with sensenet</div><br /><a style="color: #26a69a; line-height: 2rem" rel="noopener noreferrer" target="_blank" href="https://www.sensenet.com/try-it/snaas-overview">Overview</a><br /><a style="color: #26a69a; line-height: 2rem" rel="noopener noreferrer" target="_blank" href="https://community.sensenet.com/docs/getting-started">Getting started</a><br /><a rel="noopener noreferrer" target="_blank" href="https://community.sensenet.com/tutorials" style="color: #26a69a; line-height: 2rem">Tutorials</a><br /><a style="color: #26a69a; line-height: 2rem" target="_blank" href="/">Example apps</a>',
-        },
-        minWidth: {
-          default: '45%',
-        },
-      },
-      {
-        title: 'API documentation',
-        widgetType: 'markdown',
-        settings: {
-          content:
-            '<div>Discover capabilites of the API</div><br /><a style="color: #26a69a; line-height: 2rem" rel="noopener noreferrer" target="_blank" href="https://community.sensenet.com/docs/odata-rest-api/">REST API</a><br /><a style="color: #26a69a; line-height: 2rem" rel="noopener noreferrer" target="_blank" href="https://community.sensenet.com/docs/odata-rest-api/">Content Management API</a><br /><a rel="noopener noreferrer" target="_blank" href="https://community.sensenet.com/docs/built-in-odata-actions-and-functions/" style="color: #26a69a; line-height: 2rem">Document Preview API</a><br /><a style="color: #26a69a; line-height: 2rem" rel="noopener norefferrer" target="_blank" href="https://community.sensenet.com/docs/odata-rest-api/">User Management API</a>',
-        },
-        minWidth: {
-          default: '45%',
-        },
-      },
-      {
-        title: 'Have any questions?',
-        widgetType: 'markdown',
-        settings: {
-          content:
-            "<div style='text-align:center;'>If you need any help or further information, feel free to contact us!<br /><br /><a rel='noopener noreferrer' target='_blank' href='https://www.sensenet.com/contact' style='text-decoration: none;'><button style='margin-bottom: 2em' class='MuiButtonBase-root MuiButton-root MuiButton-text MuiButton-contained MuiButton-containedPrimary'>Contact us</button></a></div>",
-        },
-        minWidth: {
-          default: '100%',
-        },
-      },
-    ],
-  },
   default: {
     content: {
       browseType: 'explorer',
